Search Filter: Add "Any" option as a day selection
Users want to see all options given to them regardless of day [1].
By adding the "Any" option to the "When do you need it?" selection,
users are now able to see the full list of results.

Next, because we still want to show which venues are closed for todays
date, we filter out the ones that are closed today into the closed
markers.

Links:
1. https://github.com/code-for-chapel-hill/NC-COVID-Support/issues/107