Search Filter: Default day selection to Any
Introduced in 62b52ca ("Search Filter: Add "Any" option as a day selection
(#153)"), but the dropdown menu would still select the users current day
as the default option. Instead, we want to make the selected day default
to "Any". Therefore, set the default day selection to "Any".

Lastly, create a constant for "dayAny" so we know that it's not a valid
day selection.