Set map key to use SVG icons
Commit dd479e ("Adds business icons to leaflet markers #148 (#167)") changed
out the icons for the map to SVG based map icons. The map key icons no
longer match that on the map. We fix this by using the generator that
generates the SVG icons, and pass them to the IconListItem component.